+++ src/main/java/crelle/family/SysApplication.java
@@ -1,0 +1,15 @@
+package crelle.family;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@MapperScan(basePackages = "crelle.family.mapper")
+public class SysApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SysApplication.class, args);
+    }
+
+}
