+++ src/main/java/crelle/family/common/config/CustomFilterInvocationSecurityMetadataSource.java
@@ -1,0 +1,60 @@
+package crelle.family.common.config;
+
+import crelle.family.model.entity.Menu;
+import crelle.family.model.entity.Role;
+import crelle.family.service.MenuService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.ConfigAttribute;
+import org.springframework.security.access.SecurityConfig;
+import org.springframework.security.web.FilterInvocation;
+import org.springframework.security.web.access.intercept.FilterInvocationSecurityMetadataSource;
+import org.springframework.stereotype.Component;
+import org.springframework.util.AntPathMatcher;
+
+import java.util.Collection;
+import java.util.List;
+
+/**
+ * @author:crelle
+ * @date:2021/3/24
+ * @title:
+ * @description:这个类的作用，主要是根据用户传来的请求地址，分析出请求需要的角色
+ * @params:
+ * @return:
+ * @throw:
+ */
+@Component
+public class CustomFilterInvocationSecurityMetadataSource implements FilterInvocationSecurityMetadataSource {
+
+    @Autowired
+    MenuService menuService;
+
+    AntPathMatcher antPathMatcher = new AntPathMatcher();
+
+    @Override
+    public Collection<ConfigAttribute> getAttributes(Object object) throws IllegalArgumentException {
+        String requestUrl = ((FilterInvocation) object).getRequestUrl();
+        List<Menu> menus = menuService.queryMenus();
+        for (Menu menu : menus) {
+            if (antPathMatcher.match(menu.getUrl(), requestUrl)) {
+                List<Role> roles = menu.getRoles();
+                String[] str = new String[roles.size()];
+                for (int i = 0; i < roles.size(); i++) {
+                    str[i] = roles.get(i).getName();
+                }
+                return SecurityConfig.createList(str);
+            }
+        }
+        return SecurityConfig.createList("ROLE_LOGIN");
+    }
+
+    @Override
+    public Collection<ConfigAttribute> getAllConfigAttributes() {
+        return null;
+    }
+
+    @Override
+    public boolean supports(Class<?> clazz) {
+        return true;
+    }
+}
