+++ src/main/java/crelle/family/dao/MenuDao.java
@@ -1,0 +1,15 @@
+package crelle.family.dao;
+
+import crelle.family.model.entity.Menu;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * @author:crelle
+ * @className:MenuDao
+ * @version:1.0.0
+ * @date:2021/3/22
+ * @description:XX
+ **/
+public interface MenuDao extends JpaRepository<Menu,Long> {
+
+}
