--- src/main/java/plant/common/config/CustomUrlDecisionManager.java
@@ -1,55 +1,0 @@
-package plant.common.config;
-
-import org.springframework.security.access.AccessDecisionManager;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.access.ConfigAttribute;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
-import org.springframework.security.authentication.InsufficientAuthenticationException;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.stereotype.Component;
-
-import java.util.Collection;
-
-/**
- *@author:crelle
- *@date:2021/3/24
- *@title:
- *@description:分析用户请求的资源路径是否有权限访问
- *@params:
- *@return:
- *@throw:
- */
-@Component
-public class CustomUrlDecisionManager implements AccessDecisionManager {
-    @Override
-    public void decide(Authentication authentication, Object object, Collection<ConfigAttribute> configAttributes) throws AccessDeniedException, InsufficientAuthenticationException {
-        for (ConfigAttribute configAttribute : configAttributes) {
-            String needRole = configAttribute.getAttribute();
-            if ("ROLE_LOGIN".equals(needRole)) {
-                if (authentication instanceof AnonymousAuthenticationToken) {
-                    throw new AccessDeniedException("尚未登录，请登录!");
-                }else {
-                    return;
-                }
-            }
-            Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-            for (GrantedAuthority authority : authorities) {
-                if (authority.getAuthority().equals(needRole)) {
-                    return;
-                }
-            }
-        }
-        throw new AccessDeniedException("权限不足，请联系管理员!");
-    }
-
-    @Override
-    public boolean supports(ConfigAttribute attribute) {
-        return true;
-    }
-
-    @Override
-    public boolean supports(Class<?> clazz) {
-        return true;
-    }
-}
