--- src/main/java/plant/common/config/SwaggerConfig.java
@@ -1,39 +1,0 @@
-package plant.common.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-/**
- * @author:crelle
- * @className:SwaggerConfig
- * @version:1.0.0
- * @date:2021/3/22
- * @description:XX
- **/
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-    @Bean
-    public Docket createRestApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .pathMapping("/")
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("plant.controller"))
-                .paths(PathSelectors.any())
-                .build().apiInfo(new ApiInfoBuilder()
-                        .title("家庭百科系统接口列表")
-                        .description("作者信息:")
-                        .version("9.0")
-                        .contact(new Contact("crelle","https://www.cnblogs.com/crelle/","383362161@qq.com"))
-                        .license("The Apache License")
-                        .licenseUrl("http://www.baidu.com")
-                        .build());
-    }
-}
