--- src/main/java/plant/model/entity/Role.java
@@ -1,66 +1,0 @@
-package plant.model.entity;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-/**
- *@author:crelle
- *@date:2021/3/18
- *@title:
- *@description:
- *@params:
- *@return:
- *@throw:
- */
-@ApiModel(value = "role",description = "角色实体")
-@Entity(name = "role")
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @ApiModelProperty(value = "角色标识")
-    private Long id;
-
-    @ApiModelProperty(value = "角色编码")
-    private String name;
-
-    @ApiModelProperty(value = "角色名称")
-    private String nameZh;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getNameZh() {
-        return nameZh;
-    }
-
-    public void setNameZh(String nameZh) {
-        this.nameZh = nameZh;
-    }
-
-    @Override
-    public String toString() {
-        return "Role{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", nameZh='" + nameZh + '\'' +
-                '}';
-    }
-}
